requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
loguru>=0.7.2
SQLAlchemy>=2.0.32
asyncmy>=0.2.9
pydantic>=2.9
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
